initial setup, nothing works
Just adding some initial files and what not. This is preparing to start
writing out a build/publishing chain for this site/book. I think it should
be easy enough to scrap together some simple templating along with a Makefile
and some easy stuffs for shoving into pandoc to make nice ebooks and what-not.